Finally fix the ppu glitches on level load for good. (Well, I hope anyway)
I suspect it was related to a bug with the ppu/bit 0 bus conflicts.

Turning off interrupts while loading the level seems to have helped regardless of what the problem is, and probably wasn't a bad idea anyway.

Probably should have cleaned up the vblank_waits in this method.. will need to do that in the future.